split to two CRDs
